+++ crawler/prospector_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestProspectorInit(t *testing.T) {
 
-	fileConfig := config.FileConfig{
+	prospectorConfig := config.ProspectorConfig{
 		ScanFrequency:       "15s",
 		IgnoreOlder:         "100m",
 		HarvesterBufferSize: 100,
@@ -17,7 +17,7 @@ func TestProspectorInit(t *testing.T) {
 	}
 
 	prospector := Prospector{
-		FileConfig: fileConfig,
+		ProspectorConfig: prospectorConfig,
 	}
 
 	assert.NotNil(t, prospector)
@@ -25,74 +25,73 @@ func TestProspectorInit(t *testing.T) {
 	prospector.Init()
 
 	// Predefined values expected
-	assert.Equal(t, 100*time.Minute, prospector.FileConfig.IgnoreOlderDuration)
-	assert.Equal(t, 15*time.Second, prospector.FileConfig.ScanFrequencyDuration)
-	assert.Equal(t, 100, prospector.FileConfig.HarvesterBufferSize)
-	assert.Equal(t, true, prospector.FileConfig.TailOnRotate)
+	assert.Equal(t, 100*time.Minute, prospector.ProspectorConfig.IgnoreOlderDuration)
+	assert.Equal(t, 15*time.Second, prospector.ProspectorConfig.ScanFrequencyDuration)
+	assert.Equal(t, 100, prospector.ProspectorConfig.HarvesterBufferSize)
+	assert.Equal(t, true, prospector.ProspectorConfig.TailOnRotate)
 }
 
 func TestProspectorInitEmpty(t *testing.T) {
 
-	fileConfig := config.FileConfig{
+	prospectorConfig := config.ProspectorConfig{
 		ScanFrequency:       "",
 		IgnoreOlder:         "",
 		HarvesterBufferSize: 0,
 	}
 
 	prospector := Prospector{
-		FileConfig: fileConfig,
+		ProspectorConfig: prospectorConfig,
 	}
 
 	prospector.Init()
 
 	// Default values expected
-	assert.Equal(t, config.DefaultIgnoreOlderDuration, prospector.FileConfig.IgnoreOlderDuration)
-	assert.Equal(t, config.DefaultScanFrequency, prospector.FileConfig.ScanFrequencyDuration)
-	assert.Equal(t, config.DefaultHarvesterBufferSize, prospector.FileConfig.HarvesterBufferSize)
+	assert.Equal(t, config.DefaultIgnoreOlderDuration, prospector.ProspectorConfig.IgnoreOlderDuration)
+	assert.Equal(t, config.DefaultScanFrequency, prospector.ProspectorConfig.ScanFrequencyDuration)
 }
 
 func TestProspectorInitNotSet(t *testing.T) {
 
-	fileConfig := config.FileConfig{}
+	prospectorConfig := config.ProspectorConfig{}
 
 	prospector := Prospector{
-		FileConfig: fileConfig,
+		ProspectorConfig: prospectorConfig,
 	}
 
 	prospector.Init()
 
 	// Default values expected
-	assert.Equal(t, config.DefaultIgnoreOlderDuration, prospector.FileConfig.IgnoreOlderDuration)
-	assert.Equal(t, config.DefaultScanFrequency, prospector.FileConfig.ScanFrequencyDuration)
-	assert.Equal(t, config.DefaultHarvesterBufferSize, prospector.FileConfig.HarvesterBufferSize)
-	assert.Equal(t, config.DefaultTailOnRotate, prospector.FileConfig.TailOnRotate)
+	assert.Equal(t, config.DefaultIgnoreOlderDuration, prospector.ProspectorConfig.IgnoreOlderDuration)
+	assert.Equal(t, config.DefaultScanFrequency, prospector.ProspectorConfig.ScanFrequencyDuration)
+	assert.Equal(t, config.DefaultHarvesterBufferSize, prospector.ProspectorConfig.HarvesterBufferSize)
+	assert.Equal(t, config.DefaultTailOnRotate, prospector.ProspectorConfig.TailOnRotate)
 }
 
 func TestProspectorInitScanFrequency0(t *testing.T) {
 
-	fileConfig := config.FileConfig{
+	prospectorConfig := config.ProspectorConfig{
 		ScanFrequency: "0s",
 	}
 
 	prospector := Prospector{
-		FileConfig: fileConfig,
+		ProspectorConfig: prospectorConfig,
 	}
 
 	prospector.Init()
 
 	var zero time.Duration = 0
 	// 0 expected
-	assert.Equal(t, zero, prospector.FileConfig.ScanFrequencyDuration)
+	assert.Equal(t, zero, prospector.ProspectorConfig.ScanFrequencyDuration)
 }
 
 func TestProspectorInitInvalidScanFrequency(t *testing.T) {
 
-	fileConfig := config.FileConfig{
+	prospectorConfig := config.ProspectorConfig{
 		ScanFrequency: "abc",
 	}
 
 	prospector := Prospector{
-		FileConfig: fileConfig,
+		ProspectorConfig: prospectorConfig,
 	}
 
 	err := prospector.Init()
@@ -101,12 +100,12 @@ func TestProspectorInitInvalidScanFrequency(t *testing.T) {
 
 func TestProspectorInitInvalidIgnoreOlder(t *testing.T) {
 
-	fileConfig := config.FileConfig{
+	prospectorConfig := config.ProspectorConfig{
 		IgnoreOlder: "abc",
 	}
 
 	prospector := Prospector{
-		FileConfig: fileConfig,
+		ProspectorConfig: prospectorConfig,
 	}
 
 	err := prospector.Init()