+++ harvester/log.go
@@ -60,7 +60,7 @@ func (h *Harvester) Harvest() {
 			Offset:   h.Offset,
 			Line:     line,
 			Text:     text,
-			Fields:   &h.FileConfig.Fields,
+			Fields:   &h.ProspectorConfig.Fields,
 			Fileinfo: &info,
 		}
 		h.Offset += int64(bytesread)
@@ -217,15 +217,15 @@ func (h *Harvester) handleReadlineError(lastTimeRead time.Time, err error) error
 		// Check to see if the file was truncated
 		info, _ := h.file.Stat()
 
-		if h.FileConfig.IgnoreOlder != "" {
-			logp.Debug("harvester", "Ignore Unmodified After: %s", h.FileConfig.IgnoreOlder)
+		if h.ProspectorConfig.IgnoreOlder != "" {
+			logp.Debug("harvester", "Ignore Unmodified After: %s", h.ProspectorConfig.IgnoreOlder)
 		}
 
 		if info.Size() < h.Offset {
 			logp.Debug("harvester", "File truncated, seeking to beginning: %s", h.Path)
 			h.file.Seek(0, os.SEEK_SET)
 			h.Offset = 0
-		} else if age := time.Since(lastTimeRead); age > h.FileConfig.IgnoreOlderDuration {
+		} else if age := time.Since(lastTimeRead); age > h.ProspectorConfig.IgnoreOlderDuration {
 			// if lastTimeRead was more than ignore older and ignore older is set, this file is probably dead. Stop watching it.
 			logp.Debug("harvester", "Stopping harvest of ", h.Path, "last change was: ", age)
 			return err